Update resolveConnections to logic apps shared
